+++ bpod_core/bpod.py
@@ -1,6 +1,7 @@
-from __future__ import annotations
+"""Module for interfacing with the Bpod Finite State Machine."""
 
 import logging
+import struct
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, NamedTuple
 
@@ -9,16 +10,28 @@ from serial import SerialException
 from serial.tools.list_ports import comports
 
 from bpod_core import __version__ as bpod_core_version
-from bpod_core.serial import ExtendedSerial
+from bpod_core.com import ExtendedSerial
 
 if TYPE_CHECKING:
     from _typeshed import ReadableBuffer  # noqa: F401
 
 PROJECT_NAME = 'bpod-core'
-VIDS_BPOD = [0x16C0]  # vendor IDs of supported Bpod devices
+VENDOR_IDS_BPOD = [0x16C0]  # vendor IDs of supported Bpod devices
 MIN_BPOD_FW_VERSION = (23, 0)  # minimum supported firmware version (major, minor)
 MIN_BPOD_HW_VERSION = 3  # minimum supported hardware version
 MAX_BPOD_HW_VERSION = 4  # maximum supported hardware version
+CHANNEL_TYPES = {
+    b'U': 'Serial',
+    b'X': 'SoftCode',
+    b'Z': 'SoftCodeApp',
+    b'F': 'FlexIO',
+    b'S': 'SPI',
+    b'D': 'Digital',
+    b'B': 'BNC',
+    b'W': 'Wire',
+    b'V': 'Valve',
+    b'P': 'Port',
+}
 
 logger = logging.getLogger(__name__)
 
@@ -39,7 +52,7 @@ class HardwareConfiguration(NamedTuple):
 
     max_states: int
     """Maximum number of supported states in a single state machine description."""
-    timer_period: int
+    cycle_period: int
     """Period of the state machine's refresh cycle during a trial in microseconds."""
     max_serial_events: int
     """Maximum number of behavior events allocatable among connected modules."""
@@ -60,6 +73,16 @@ class HardwareConfiguration(NamedTuple):
     output_description: bytes
     """Array indicating the state machine's onboard output channel types."""
 
+    @property
+    def cycle_frequency(self) -> int:
+        """Frequency of the state machine's refresh cycle during a trial in Hertz."""
+        return 1000000 // self.cycle_period
+
+    @property
+    def n_modules(self) -> int:
+        """Number of modules supported by the state machine."""
+        return self.input_description.count(b'U')
+
 
 class BpodError(Exception):
     """
@@ -74,20 +97,22 @@ class Bpod:
     """Bpod class for interfacing with the Bpod Finite State Machine."""
 
     _version: VersionInfo
-    _hardware_config: HardwareConfiguration
+    _hardware: HardwareConfiguration
     serial0: ExtendedSerial
     """Primary serial device for communication with the Bpod."""
-    serial1: ExtendedSerial
+    serial1: ExtendedSerial | None = None
     """Secondary serial device for communication with the Bpod."""
     serial2: ExtendedSerial | None = None
     """Tertiary serial device for communication with the Bpod - used by Bpod 2+ only."""
+    inputs: NamedTuple
+    outputs: NamedTuple
 
     @validate_call
     def __init__(self, port: str | None = None, serial_number: str | None = None):
         logger.info(f'bpod_core {bpod_core_version}')
 
         # identify Bpod by port or serial number
-        port, serial_number = self._identify_bpod(port, serial_number)
+        port, self._serial_number = self._identify_bpod(port, serial_number)
 
         # open primary serial port
         self.serial0 = ExtendedSerial()
@@ -101,7 +126,7 @@ class Bpod:
         self._get_hardware_configuration()
 
         # configure input and output channels
-        self._configure_channels()
+        self._configure_io()
 
         # detect additional serial ports
         self._detect_additional_serial_ports()
@@ -111,7 +136,7 @@ class Bpod:
         logger.info(f'Connected to Bpod Finite State Machine {machine} on {self.port}')
         logger.info(
             f'Firmware Version {"{}.{}".format(*self.version.firmware)}, '
-            f'Serial Number {serial_number}, PCB Revision {self.version.pcb}'
+            f'Serial Number {self._serial_number}, PCB Revision {self.version.pcb}'
         )
 
     def __enter__(self):
@@ -125,9 +150,41 @@ class Bpod:
     def __del__(self):
         self.close()
 
-    @staticmethod
+    def _sends_discovery_byte(
+        self,
+        port: str,
+        byte: bytes = b'\xde',
+        timeout: float = 0.11,
+        trigger: bytes | None = None,
+    ) -> bool:
+        r"""Check if the device on the given port sends a discovery byte.
+
+        Parameters
+        ----------
+        port : str
+            The name of the serial port to check (e.g., '/dev/ttyUSB0' or 'COM3').
+        byte : bytes, optional
+            The discovery byte to expect from the device. Defaults to b'\\xde'.
+        timeout : float, optional
+            Timeout period (in seconds) for the serial read operation. Defaults to 0.11.
+        trigger : bytes, optional
+            An optional command to send on serial0 before reading from the given device.
+
+        Returns
+        -------
+        bool
+            Whether the given device responded with the expected discovery byte or not.
+        """
+        try:
+            with ExtendedSerial(port, timeout=timeout) as ser:
+                if trigger is not None and getattr(self, 'serial0', None) is not None:
+                    self.serial0.write(trigger)
+                return ser.read(1) == byte
+        except SerialException:
+            return False
+
     def _identify_bpod(
-        port: str | None = None, serial_number: str | None = None
+        self, port: str | None = None, serial_number: str | None = None
     ) -> tuple[str, str | None]:
         """
         Try to identify a supported Bpod based on port or serial number.
@@ -154,23 +211,14 @@ class Bpod:
         BpodError
             If no Bpod is found or the indicated device is not supported.
         """
-
-        def sends_discovery_byte(port: str) -> bool:
-            """Check if the device on the given port sends a discovery byte."""
-            try:
-                with ExtendedSerial(port, timeout=0.15) as ser:
-                    return ser.read(1) == bytes([222])
-            except SerialException:
-                return False
-
         # If no port or serial number provided, try to automagically find an idle Bpod
         if port is None and serial_number is None:
             try:
                 port_info = next(
                     p
                     for p in comports()
-                    if getattr(p, 'vid', None) in VIDS_BPOD
-                    and sends_discovery_byte(p.device)
+                    if getattr(p, 'vid', None) in VENDOR_IDS_BPOD
+                    and self._sends_discovery_byte(p.device)
                 )
             except StopIteration as e:
                 raise BpodError('No available Bpod found') from e
@@ -183,7 +231,7 @@ class Bpod:
                     p
                     for p in comports()
                     if p.serial_number == serial_number
-                    and sends_discovery_byte(p.device)
+                    and self._sends_discovery_byte(p.device)
                 )
             except (StopIteration, AttributeError) as e:
                 raise BpodError(f'No device with serial number {serial_number}') from e
@@ -195,7 +243,7 @@ class Bpod:
             except (StopIteration, AttributeError) as e:
                 raise BpodError(f'Port not found: {port}') from e
 
-        if port_info.vid not in VIDS_BPOD:
+        if port_info.vid not in VENDOR_IDS_BPOD:
             raise BpodError('Device is not a supported Bpod')
         return port_info.device, port_info.serial_number
 
@@ -213,6 +261,7 @@ class Bpod:
         BpodError
             If the hardware version or firmware version is not supported.
         """
+        logger.debug('Retrieving version information')
         v_major, machine_type = self.serial0.query_struct(b'F', '<2H')
         v_minor = self.serial0.query_struct(b'f', '<H')[0] if v_major > 22 else 0
         v_firmware = (v_major, v_minor)
@@ -231,6 +280,7 @@ class Bpod:
 
     def _get_hardware_configuration(self) -> None:
         """Retrieve the Bpod's onboard hardware configuration."""
+        logger.debug('Retrieving onboard hardware configuration')
         if self.version.firmware > (22, 0):
             hardware_conf = list(self.serial0.query_struct(b'H', '<2H6B'))
         else:
@@ -238,80 +288,76 @@ class Bpod:
             hardware_conf.insert(-4, 3)  # max bytes per serial msg always = 3
         hardware_conf.extend(self.serial0.read_struct(f'<{hardware_conf[-1]}s1B'))
         hardware_conf.extend(self.serial0.read_struct(f'<{hardware_conf[-1]}s'))
-        self._hardware_config = HardwareConfiguration(*hardware_conf)
+        self._hardware = HardwareConfiguration(*hardware_conf)
 
-    def _configure_channels(self) -> None:
-        def collect_channels(description: bytes, dictionary: dict, channel_cls: type):
-            """
-            Generate a collection of Bpod channels.
-
-            This method takes a channel description array (as returned by the Bpod), a
-            dictionary mapping keys to names, and a channel class. It generates named
-            tuple instances and sets them as attributes on the current Bpod instance.
-            """
+    def _configure_io(self) -> None:
+        """Configure the input and output channels of the Bpod."""
+        logger.debug('Configuring I/O')
+        for description, channel_class in (
+            (self._hardware.input_description, Input),
+            (self._hardware.output_description, Output),
+        ):
+            n_channels = len(description)
+            io_class = f'{channel_class.__name__.lower()}s'
             channels = []
             types = []
 
-            for idx in range(len(description)):
-                io_key = description[idx : idx + 1]
-                if bytes(io_key) in dictionary:
-                    n = description[:idx].count(io_key) + 1
-                    name = f'{dictionary[io_key]}{n}'
-                    channels.append(channel_cls(self, name, io_key, idx))
-                    types.append((name, channel_cls))
+            # loop over the description array and create channels
+            for idx, io_key in enumerate(struct.unpack(f'<{n_channels}c', description)):
+                if io_key not in CHANNEL_TYPES:
+                    raise RuntimeError(f'Unknown {io_class[:-1]} type: {io_key}')
+                n = description[:idx].count(io_key) + 1
+                name = f'{CHANNEL_TYPES[io_key]}{n}'
+                channels.append(channel_class(self, name, io_key, idx))
+                types.append((name, channel_class))
 
-            cls_name = f'{channel_cls.__name__.lower()}s'
-            setattr(self, cls_name, NamedTuple(cls_name, types)._make(channels))
+            # store channels to NamedTuple and set the latter as a class attribute
+            named_tuple = NamedTuple(io_class, types)._make(channels)
+            setattr(self, io_class, named_tuple)
 
-        logger.debug('Configuring I/O ports')
-        input_dict = {b'B': 'BNC', b'V': 'Valve', b'P': 'Port', b'W': 'Wire'}
-        output_dict = {b'B': 'BNC', b'V': 'Valve', b'P': 'PWM', b'W': 'Wire'}
-        collect_channels(self._hardware_config.input_description, input_dict, Input)
-        collect_channels(self._hardware_config.output_description, output_dict, Output)
+        # set the enabled state of the input channels
+        self._set_enable_inputs()
 
     def _detect_additional_serial_ports(self) -> None:
         """Detect additional USB-serial ports."""
+        logger.debug('Detecting additional USB-serial ports')
+
         # First, assemble a list of candidate ports
         candidate_ports = [
-            p.device for p in comports() if p.vid in VIDS_BPOD and p.device != self.port
+            p.device
+            for p in comports()
+            if p.serial_number == self._serial_number and p.device != self.port
         ]
 
         # Exclude those devices from the list that are already sending a discovery byte
-        for port in candidate_ports:
-            try:
-                with ExtendedSerial(port, timeout=0.15) as ser:
-                    if ser.read(1) == bytes([222]):
-                        candidate_ports.remove(port)
-            except SerialException:
-                pass
+        # NB: this should not be necessary, as we already filter for devices with
+        #     identical USB serial number.
+        # for port in candidate_ports:
+        #     if self._sends_discovery_byte(port):
+        #         candidate_ports.remove(port)
 
-        # Find second USB-serial port
-        for port in candidate_ports:
-            try:
-                with ExtendedSerial(port, timeout=0.15) as ser:
-                    self.serial0.write(b'{')
-                    if ser.read(1) == bytes([222]):
-                        ser.reset_input_buffer()
-                        ser.timeout = None
-                        self.serial1 = ser
-                        candidate_ports.remove(port)
-                        break
-            except SerialException:
-                pass
+        # Find secondary USB-serial port
+        if self._version.firmware >= (23, 0):
+            for port in candidate_ports:
+                if self._sends_discovery_byte(port, bytes([222]), trigger=b'{'):
+                    self.serial1 = ExtendedSerial()
+                    self.serial1.port = port
+                    candidate_ports.remove(port)
+                    logger.debug(f'Detected secondary USB-serial port: {port}')
+                    break
+            if self.serial2 is None:
+                raise BpodError('Could not detect secondary serial port')
 
-        # State Machine 2+ uses a third USB-serial port
+        # State Machine 2+ uses a third USB-serial port for FlexIO
         if self.version.machine == 4:
             for port in candidate_ports:
-                try:
-                    with ExtendedSerial(port, timeout=0.15) as ser:
-                        self.serial0.write(b'}')
-                        if ser.read(1) == bytes([223]):
-                            ser.reset_input_buffer()
-                            ser.timeout = None
-                            self.serial2 = ser
-                            break
-                except SerialException:
-                    pass
+                if self._sends_discovery_byte(port, bytes([223]), trigger=b'}'):
+                    self.serial2 = ExtendedSerial()
+                    self.serial2.port = port
+                    logger.debug(f'Detected tertiary USB-serial port: {port}')
+                    break
+            if self.serial2 is None:
+                raise BpodError('Could not detect tertiary serial port')
 
     def _handshake(self):
         """
@@ -332,6 +378,23 @@ class Bpod:
         finally:
             self.serial0.reset_input_buffer()
         logger.debug(f'Handshake with Bpod on {self.port} successful')
+
+    def _test_psram(self) -> bool:
+        """
+        Test the Bpod's PSRAM.
+
+        Returns
+        -------
+        bool
+            True if the PSRAM test passed, False otherwise.
+        """
+        return self.serial0.verify(b'_')
+
+    def _set_enable_inputs(self) -> bool:
+        logger.debug('Updating enabled state of input channels')
+        enable = [i.enabled for i in self.inputs]
+        self.serial0.write_struct(f'<c{self._hardware.n_inputs}?', b'E', *enable)
+        return self.serial0.read(1) == b'\x01'
 
     @property
     def port(self) -> str | None:
@@ -365,37 +428,52 @@ class Bpod:
             self.serial0.write(b'Z')
             self.serial0.close()
 
+    def set_status_led(self, enabled: bool) -> bool:
+        """
+        Enable or disable the Bpod's status LED.
+
+        Parameters
+        ----------
+        enabled : bool
+            True to enable the status LED, False to disable.
+
+        Returns
+        -------
+        bool
+            True if the operation was successful, False otherwise.
+        """
+        self.serial0.write_struct('<c?', b':', enabled)
+        return self.serial0.verify(b'_')
+
     def update_modules(self):
-        self.serial0.write(b'M')
-        # modules = []
-        # for i in range(len(modules)):
-        #     if self.serial0.read() == bytes([1]):
-        #         continue
-        #     firmware_version = self.serial0.read(4, np.uint32)[0]
-        #     name = self.read(int(self.serial0.read())).decode('utf-8')
-        #     port = i + 1
-        #     m = Module()
-        #     while self.serial0.read() == b'\x01':
-        #         match self.serial0.read():
-        #             case b'#':
-        #                 number_of_events = self.serial0.read(1, np.uint8)[0]
-        #             case b'E':
-        #                 for event_index in range(self.serial0.read(1, np.uint8)[0]):
-        #                     l_event_name = self.serial0.read(1, np.uint8)[0]
-        #                     module['events']['index'] = event_index
-        #                     module['events']['name'] = self.serial0.read(
-        #                         l_event_name, str
-        #                     )[0]
-        #         modules[i] = module
-        #     self._children = modules
         pass
+        # self.serial0.write(b'M')
+        # modules = []
+        # for idx in range(self._hardware.n_modules):
+        #     is_connected = self.serial0.read_struct('<?')[0]
+        #     if not is_connected:
+        #         module_name = f'{CHANNEL_TYPES[b"U"]}{idx + 1}'
+        #     else:
+        #         firmware_version, n = self.serial0.read_struct('<IB')
+        #         module_name, more_info = self.serial0.read_struct(f'<{n}s?')
+        #         while more_info:
+        #             match self.serial0.read(1):
+        #                 case b'#':
+        #                     n_serial_events = self.serial0.read_struct('<B')[0]
+        #                 case b'E':
+        #                     n_evt_names = self.serial0.read(0)
+        #                     evt_names = []
+        #                     for i_event_name in range(n_evt_names):
+        #                         n = self.serial0.read_struct('<B')[0]
+        #                         evt_names.append(self.serial0.read_struct(f'<{n}s'))
+        #             more_info = self.serial0.read_struct('<?')[0]
 
 
 class Channel(ABC):
     """Abstract base class representing a channel on the Bpod device."""
 
     @abstractmethod
-    def __init__(self, bpod: Bpod, name: str, io_type: bytes, index: int):
+    def __init__(self, bpod: Bpod, name: str, io_key: bytes, index: int):
         """
         Abstract base class representing a channel on the Bpod device.
 
@@ -405,17 +483,15 @@ class Channel(ABC):
             The Bpod instance associated with the channel.
         name : str
             The name of the channel.
-        io_type : bytes
-            The I/O type of the channel (e.g., 'B', 'V', 'P').
+        io_key : bytes
+            The I/O type of the channel (e.g., b'B', b'V', b'P').
         index : int
             The index of the channel.
         """
         self.name = name
-        self.io_type = io_type
+        self.io_type = io_key
         self.index = index
-        self._query = bpod.serial0.query
-        self._write = bpod.serial0.write
-        self._validate_response = bpod.serial0.verify
+        self._serial0 = bpod.serial0
 
     def __repr__(self):
         return self.__class__.__name__ + '()'
@@ -424,16 +500,24 @@ class Channel(ABC):
 class Input(Channel):
     """Input channel class representing a digital input channel."""
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, bpod: Bpod, name: str, io_key: bytes, index: int):
         """
         Input channel class representing a digital input channel.
 
         Parameters
         ----------
-        *args, **kwargs
-            Arguments to be passed to the base class constructor.
+        bpod : Bpod
+            The Bpod instance associated with the channel.
+        name : str
+            The name of the channel.
+        io_key : bytes
+            The I/O type of the channel (e.g., b'B', b'V', b'P').
+        index : int
+            The index of the channel.
         """
-        super().__init__(*args, **kwargs)
+        super().__init__(bpod, name, io_key, index)
+        self._set_enable_inputs = bpod._set_enable_inputs
+        self._enabled = io_key in (b'PBWF')  # Enable Port, BNC, Wire and FlexIO inputs
 
     def read(self) -> bool:
         """
@@ -444,8 +528,7 @@ class Input(Channel):
         bool
             True if the input channel is active, False otherwise.
         """
-        return self._validate_response(b'I' + bytes([self.index]), b'\x01')
-        # return self._query(['I', self.index], 1) == b'\x01'
+        return self._serial0.verify([b'I', self.index])
 
     def override(self, state: bool) -> None:
         """
@@ -456,33 +539,70 @@ class Input(Channel):
         state : bool
             The state to set for the input channel.
         """
-        self._write([b'V', state])
+        self._serial0.write_struct('<cB', b'V', state)
 
-    def enable(self, state: bool) -> None:
+    def enable(self, enabled: bool) -> bool:
         """
         Enable or disable the input channel.
 
         Parameters
         ----------
-        state : bool
+        enabled : bool
+            True to enable the input channel, False to disable.
+
+        Returns
+        -------
+        bool
+            True if the operation was success, False otherwise.
+        """
+        self._enabled = enabled
+        success = self._set_enable_inputs()
+        return success
+
+    @property
+    def enabled(self) -> bool:
+        """
+        Check if the input channel is enabled.
+
+        Returns
+        -------
+        bool
+            True if the input channel is enabled, False otherwise.
+        """
+        return self._enabled
+
+    @enabled.setter
+    def enabled(self, enabled: bool) -> None:
+        """
+        Enable or disable the input channel.
+
+        Parameters
+        ----------
+        enabled : bool
             True to enable the input channel, False to disable.
         """
-        pass
+        self.enable(enabled)
 
 
 class Output(Channel):
     """Output channel class representing a digital output channel."""
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, bpod: Bpod, name: str, io_key: bytes, index: int):
         """
         Output channel class representing a digital output channel.
 
         Parameters
         ----------
-        *args, **kwargs
-            Arguments to be passed to the base class constructor.
+        bpod : Bpod
+            The Bpod instance associated with the channel.
+        name : str
+            The name of the channel.
+        io_key : bytes
+            The I/O type of the channel (e.g., b'B', b'V', b'P').
+        index : int
+            The index of the channel.
         """
-        super().__init__(*args, **kwargs)
+        super().__init__(bpod, name, io_key, index)
 
     def override(self, state: bool | int) -> None:
         """
@@ -496,4 +616,23 @@ class Output(Channel):
         """
         if isinstance(state, int) and self.io_type in (b'D', b'B', b'W'):
             state = state > 0
-        self._write([b'O', self.index, bytes([state])])
+        self._serial0.write_struct('<c2B', b'O', self.index, state)
+
+
+# class Module:
+#     """Base class for Bpod modules."""
+#
+#     def __init__(
+#         self,
+#         bpod: Bpod,
+#         index: int,
+#         name: str,
+#         n_events: int,
+#         event_names: list[str] | None = None,
+#         firmware_version: int | None = None,
+#         connected: bool = False,
+#     ):
+#         self._bpod = bpod
+#         self._index = index
+#         self._name = name
+#         self._n_events = n_events