+++ CHANGELOG.md
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-alpha.2]
+
+### Added
+
+- Pydantic model for State Machine
+
+### Changed
+
+- Further refactoring of the code
+- Renamed `serial` module to `com` to avoid confusion with PySerial.
+ 
 ## [0.1.0-alpha.1]
 
 ### Changed